First Commit
These two folders have the a reforcement and unfinished JS and ReactJS practices, but since these are related to ReacJS. Course that it has been already reviewed or at least in process, so there was no need to finish in those practices. Second folder is just a test project to validate the use of device emulators.